+++ cmd/server/report.go
@@ -28,21 +28,9 @@ func handleReport(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
 		return
 	}
-	apiKey := strings.Replace(r.Header.Get("Authorization"), "token ", "", 1)
-	if len(apiKey) == 0 {
+	if !validateAPIKey(r, false) {
 		http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
 		return
-	}
-	if ok, _, err := database.ValidateAPIKey(apiKey); !ok {
-		if err != nil {
-			log.Printf("failed to validate api key: %v", err)
-			http.Error(w, "Database unavailable", http.StatusInternalServerError)
-			return
-		}
-		if !ok {
-			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
-			return
-		}
 	}
 	body, err := ioutil.ReadAll(r.Body)
 	if err != nil {
